+++ rust/src/prediction/mod.rs
@@ -1,3 +1,4 @@
+pub mod ml;
 use crate::thermal::ThermalEngine;
 
 #[derive(Debug, Clone, PartialEq)]